feat(fundamental): return all scores in addition for sbom advisories
