+++ modules/fundamental/src/common/mod.rs
@@ -6,6 +6,7 @@ use trustify_entity::sbom_package_license::LicenseCategory;
 use utoipa::ToSchema;
 
 pub mod license_filtering;
+pub mod model;
 pub mod service;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, ToSchema, FromQueryResult)]
