+++ modules/fundamental/src/common/model/mod.rs
@@ -1,0 +1,2 @@
+mod score;
+pub use score::*;
