+++ modules/fundamental/src/vulnerability/service/mod.rs
@@ -1,22 +1,21 @@
 #[cfg(test)]
 mod test;
 
-use crate::vulnerability::model::AnalysisResult;
 use crate::{
     Error,
     advisory::model::AdvisoryHead,
+    common::model::Score,
     vulnerability::model::{
-        AnalysisAdvisory, AnalysisDetails, AnalysisResponse, Score, ScoreType,
-        VulnerabilityDetails, VulnerabilityHead, VulnerabilitySummary,
+        AnalysisAdvisory, AnalysisDetails, AnalysisResponse, AnalysisResult, VulnerabilityDetails,
+        VulnerabilityHead, VulnerabilitySummary,
     },
 };
 use futures_util::{TryFutureExt, TryStreamExt, future};
 use sea_orm::{
     EntityTrait, FromQueryResult, Statement, StreamTrait, TryGetableFromJson, prelude::*,
 };
-use std::collections::btree_map::Entry;
 use std::{
-    collections::{BTreeMap, HashMap},
+    collections::{BTreeMap, HashMap, btree_map::Entry},
     str::FromStr,
 };
 use tracing::instrument;
@@ -300,21 +299,8 @@ GROUP BY
             .await?
             .map_err(Error::from)
             .try_filter_map(async |row| {
-                // map to V3* type
-                let r#type = match row.minor_version {
-                    0 => ScoreType::V3,
-                    1 => ScoreType::V3_1,
-                    _ => return Ok(None),
-                };
-
-                Ok(Some((
-                    row.advisory_id,
-                    Score {
-                        r#type,
-                        value: row.score,
-                        severity: row.score.into(),
-                    },
-                )))
+                let advisory_id = row.advisory_id;
+                Ok(Score::try_from(row).ok().map(|score| (advisory_id, score)))
             })
             .try_fold(
                 HashMap::<Uuid, Vec<Score>>::new(),
