+++ openapi.yaml
@@ -5087,6 +5087,7 @@ components:
         - average_score
         - status
         - packages
+        - scores
         properties:
           average_score:
             type: number
@@ -5101,6 +5102,10 @@ components:
             type: array
             items:
               $ref: '#/components/schemas/SbomPackage'
+          scores:
+            type: array
+            items:
+              $ref: '#/components/schemas/Score'
           status:
             type: string
     SbomSummary:
